+++ src/exercises/exercises.ts
@@ -55,9 +55,7 @@ import { derivativeNumberReading } from './geometry/cartesian/derivativeNumberRe
 import { leadingCoefficient } from './geometry/cartesian/leadingCoefficient';
 import { midpoint } from './geometry/cartesian/midpoint';
 import { pythagore } from './geometry/euclidean/pythagore';
-import { pythagoreC } from './geometry/euclidean/pythagore copy';
 import { pythagoreCalcul } from './geometry/euclidean/pythagoreCalcul';
-import { pythagoreCalculC } from './geometry/euclidean/pythagoreCalcul copy';
 import { thales } from './geometry/euclidean/thales';
 import { thalesCalcul } from './geometry/euclidean/thalesCalcul';
 import { trigonometry } from './geometry/euclidean/trigonometry';
@@ -216,8 +214,6 @@ export const exercises = [
   thalesCalcul,
   pythagore,
   pythagoreCalcul,
-  pythagoreC,
-  pythagoreCalculC,
   trigonometry,
   trigonometryAngleCalcul,
   trigonometrySideCalcul,
